lambda/trigger_sf/requirements.txt
google-cloud-pubsub==2.31.1
grpcio==1.74.0 # not clear why but we need this dependency here for aws lambda to work
boto3==1.40.20
